requirements.txt
attrs==21.2.0
autopep8==1.5.7
boto3==1.18.21
botocore==1.21.21
botostubs==0.14.1.18.20
fuzzysearch==0.7.3
icdiff==2.0.4
iniconfig==1.1.1
jmespath==0.10.0
mypy==0.910
mypy-extensions==0.4.3
packaging==21.0
pluggy==0.13.1
pprintpp==0.4.0
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pyperclip==1.8.2
pytest==6.2.4
pytest-icdiff==0.5
python-dateutil==2.8.2
s3transfer==0.5.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
